Transition templates
Adjusted files to be in accorance with how Chris had it initially. These includes:
- HHode.m

Added transition templates to the code and added files for simpliciy when working on gen_templates.m script. This includes:
- gen_templates.m
- gen_templates_HHode.m
- templates/tempate_test2.mat

gen_templates_HHode.m was made to work with only gen_templates.m as HHode.m was modified by chris to work with HHsim.m

templates/template_test2.mat is another test file that gen_templates will export the sample template struct that is being tested